requirements.txt
crewai
streamlit
yfinance
matplotlib
langchain-groq
crewai-tools
python-dotenv